pyproject.toml
[project]
name = "rag-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "datasets>=4.5.0",
    "fastapi>=0.128.2",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "llama-cpp-python>=0.3.16",
    "openai>=2.17.0",
    "pdfplumber>=0.11.9",
    "pgvector>=0.4.2",
    "pinecone[grpc]>=8.0.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
    "uvicorn[standard]>=0.40.0",
]
